Fix: Rename client conponent to keep convention
